# api/openapi/kinopoisk/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kinopoisk Unofficial API",
    "description": "Kinopoisk Unofficial API предоставляет доступ к данным сайта https://www.kinopoisk.ru. Для доступа вы должны получить <b>токен</b>, который будет доступен после регистрации на <a href=\"https://kinopoiskapiunofficial.tech\">https://kinopoiskapiunofficial.tech</a> </br> <b>Ограничения:</b>  лимитов на общее кол-во описаны тут https://kinopoiskapiunofficial.tech/rates. Так же есть лимиты на кол-во запросов в секунду. </br> Каждый пользователь имеет ограничение в <b>20 req/sec</b>. </br> <b>Некоторые</b> эндпоинты имеют свои собственные ограничения, они указаны в описании для статуса <b>429</b>. </br>",
    "contact": {
      "name": "Support",
      "email": "support@kinopoiskapiunofficial.tech"
    },
    "version": "2.0.1"
  },
  "servers": [
    {
      "url": "https://kinopoiskapiunofficial.tech",
      "description": "Main (production) server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v2.2/films/premieres": {
      "get": {
        "tags": [
          "films"
        ],
        "summary": "получить список кинопремьер",
        "description": "Данный эндпоинт возвращает список кинопремьер. Например https://www.kinopoisk.ru/premiere/",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "description": "год релиза",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "month",
            "in": "query",
            "description": "месяц релиза",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "enum": [
                "JANUARY",
                "FEBRUARY",
                "MARCH",
                "APRIL",
                "MAY",
                "JUNE",
                "JULY",
                "AUGUST",
                "SEPTEMBER",
                "OCTOBER",
                "NOVEMBER",
                "DECEMBER"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PremiereResponse"
                }
              }
            }
          },
          "401": {
            "description": "Пустой или неправильный токен"
          },
          "402": {
            "description": "Превышен лимит запросов(или дневной, или общий)"
          },
          "429": {
            "description": "Слишком много запросов. Лимит 5 запросов в секунду"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Country": {
        "required": [
          "country"
        ],
        "type": "object",
        "properties": {
          "country": {
            "type": "string",
            "example": "США"
          }
        }
      },
      "Genre": {
        "required": [
          "genre"
        ],
        "type": "object",
        "properties": {
          "genre": {
            "type": "string",
            "example": "фантастика"
          }
        }
      },
...
      "PremiereResponse": {
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 34
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PremiereResponseItem"
            }
          }
        }
      },
      "PremiereResponseItem": {
        "required": [
          "countries",
          "duration",
          "genres",
          "kinopoiskId",
          "nameEn",
          "nameRu",
          "posterUrl",
          "posterUrlPreview",
          "premiereRu",
          "year"
        ],
        "type": "object",
        "properties": {
          "kinopoiskId": {
            "type": "integer",
            "example": 301
          },
          "nameRu": {
            "type": "string",
            "nullable": true,
            "example": "Дитя погоды"
          },
          "nameEn": {
            "type": "string",
            "nullable": true,
            "example": "Tenki no ko"
          },
          "year": {
            "type": "integer",
            "example": 2019
          },
          "posterUrl": {
            "type": "string",
            "example": "http://kinopoiskapiunofficial.tech/images/posters/kp/1219417.jpg"
          },
          "posterUrlPreview": {
            "type": "string",
            "example": "https://kinopoiskapiunofficial.tech/images/posters/kp_small/301.jpg"
          },
          "countries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Country"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "duration": {
            "type": "integer",
            "nullable": true,
            "example": 112
          },
          "premiereRu": {
            "type": "string",
            "example": "2020-06-01"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "API key to authorize requests.",
        "name": "X-API-KEY",
        "in": "header"
      }
    }
  }
}